requirements.txt
fastapi
uvicorn
pillow
pytesseract
pyzbar
python-multipart
